refactor: promote LeRobot policy module
