+++ pyproject.toml
@@ -628,7 +628,7 @@ strict = true
 warn_unused_ignores = false
 untyped_calls_exclude = ["zenoh"]
 explicit_package_bases = true
-exclude = "^dimos/models/Detic(/|$)|^dimos/experimental/robot_policy/lerobot/python/|.*/test_.|.*/tool_.|.*/conftest.py*"
+exclude = "^dimos/models/Detic(/|$)|^dimos/imitation/policy/lerobot/python/|.*/test_.|.*/tool_.|.*/conftest.py*"
 
 [[tool.mypy.overrides]]
 module = [
@@ -743,7 +743,7 @@ exclude_also = [
 max_size_kb = 75
 ignore = [
     "uv.lock",
-    "examples/external_python_module/python/uv.lock",
+    "*/uv.lock",
     "*/package-lock.json",
     "Cargo.lock",
     "*/Cargo.lock",
